+++ packages/astro/src/transitions/router.ts
@@ -1,5 +1,13 @@
 import type { TransitionBeforePreparationEvent, TransitionBeforeSwapEvent } from './events.js';
 import { TRANSITION_AFTER_SWAP, doPreparation, doSwap } from './events.js';
+import {
+	deselectScripts,
+	restoreFocus,
+	saveFocus,
+	swapBodyElement,
+	swapHeadElements,
+	swapRootAttributes,
+} from './swap-functions.js';
 import type { Direction, Fallback, Options } from './types.js';
 
 type State = {
@@ -264,139 +272,6 @@ async function updateDOM(
 	historyState?: State,
 	fallback?: Fallback
 ) {
-	// Check for a head element that should persist and returns it,
-	// either because it has the data attribute or is a link el.
-	// Returns null if the element is not part of the new head, undefined if it should be left alone.
-	const persistedHeadElement = (el: HTMLElement, newDoc: Document): Element | null => {
-		const id = el.getAttribute(PERSIST_ATTR);
-		const newEl = id && newDoc.head.querySelector(`[${PERSIST_ATTR}="${id}"]`);
-		if (newEl) {
-			return newEl;
-		}
-		if (el.matches('link[rel=stylesheet]')) {
-			const href = el.getAttribute('href');
-			return newDoc.head.querySelector(`link[rel=stylesheet][href="${href}"]`);
-		}
-		return null;
-	};
-
-	type SavedFocus = {
-		activeElement: HTMLElement | null;
-		start?: number | null;
-		end?: number | null;
-	};
-
-	const saveFocus = (): SavedFocus => {
-		const activeElement = document.activeElement as HTMLElement;
-		// The element that currently has the focus is part of a DOM tree
-		// that will survive the transition to the new document.
-		// Save the element and the cursor position
-		if (activeElement?.closest(`[${PERSIST_ATTR}]`)) {
-			if (
-				activeElement instanceof HTMLInputElement ||
-				activeElement instanceof HTMLTextAreaElement
-			) {
-				const start = activeElement.selectionStart;
-				const end = activeElement.selectionEnd;
-				return { activeElement, start, end };
-			}
-			return { activeElement };
-		} else {
-			return { activeElement: null };
-		}
-	};
-
-	const restoreFocus = ({ activeElement, start, end }: SavedFocus) => {
-		if (activeElement) {
-			activeElement.focus();
-			if (
-				activeElement instanceof HTMLInputElement ||
-				activeElement instanceof HTMLTextAreaElement
-			) {
-				if (typeof start === 'number') activeElement.selectionStart = start;
-				if (typeof end === 'number') activeElement.selectionEnd = end;
-			}
-		}
-	};
-
-	const shouldCopyProps = (el: HTMLElement): boolean => {
-		const persistProps = el.dataset.astroTransitionPersistProps;
-		return persistProps == null || persistProps === 'false';
-	};
-
-	const defaultSwap = (beforeSwapEvent: TransitionBeforeSwapEvent) => {
-		// swap attributes of the html element
-		// - delete all attributes from the current document
-		// - insert all attributes from doc
-		// - reinsert all original attributes that are named 'data-astro-*'
-		const html = document.documentElement;
-		const astroAttributes = [...html.attributes].filter(
-			({ name }) => (html.removeAttribute(name), name.startsWith('data-astro-'))
-		);
-		[...beforeSwapEvent.newDocument.documentElement.attributes, ...astroAttributes].forEach(
-			({ name, value }) => html.setAttribute(name, value)
-		);
-
-		// Replace scripts in both the head and body.
-		for (const s1 of document.scripts) {
-			for (const s2 of beforeSwapEvent.newDocument.scripts) {
-				if (
-					// Check if the script should be rerun regardless of it being the same
-					!s2.hasAttribute('data-astro-rerun') &&
-					// Inline
-					((!s1.src && s1.textContent === s2.textContent) ||
-						// External
-						(s1.src && s1.type === s2.type && s1.src === s2.src))
-				) {
-					// the old script is in the new document and doesn't have the rerun attribute
-					// we mark it as executed to prevent re-execution
-					s2.dataset.astroExec = '';
-					break;
-				}
-			}
-		}
-
-		// Swap head
-		for (const el of Array.from(document.head.children)) {
-			const newEl = persistedHeadElement(el as HTMLElement, beforeSwapEvent.newDocument);
-			// If the element exists in the document already, remove it
-			// from the new document and leave the current node alone
-			if (newEl) {
-				newEl.remove();
-			} else {
-				// Otherwise remove the element in the head. It doesn't exist in the new page.
-				el.remove();
-			}
-		}
-
-		// Everything left in the new head is new, append it all.
-		document.head.append(...beforeSwapEvent.newDocument.head.children);
-
-		// Persist elements in the existing body
-		const oldBody = document.body;
-
-		const savedFocus = saveFocus();
-
-		// this will reset scroll Position
-		document.body.replaceWith(beforeSwapEvent.newDocument.body);
-
-		for (const el of oldBody.querySelectorAll(`[${PERSIST_ATTR}]`)) {
-			const id = el.getAttribute(PERSIST_ATTR);
-			const newEl = document.querySelector(`[${PERSIST_ATTR}="${id}"]`);
-			if (newEl) {
-				// The element exists in the new page, replace it with the element
-				// from the old page so that state is preserved.
-				newEl.replaceWith(el);
-				// For islands, copy over the props to allow them to re-render
-				if (newEl.localName === 'astro-island' && shouldCopyProps(el as HTMLElement)) {
-					el.setAttribute('ssr', '');
-					el.setAttribute('props', newEl.getAttribute('props')!);
-				}
-			}
-		}
-		restoreFocus(savedFocus);
-	};
-
 	async function animate(phase: string) {
 		function isInfinite(animation: Animation) {
 			const effect = animation.effect;
@@ -430,7 +305,7 @@ async function updateDOM(
 	}
 
 	const pageTitleForBrowserHistory = document.title; // document.title will be overridden by swap()
-	const swapEvent = doSwap(preparationEvent, currentTransition.viewTransition!, defaultSwap);
+	const swapEvent = doSwap(preparationEvent, currentTransition.viewTransition!);
 	moveToLocation(swapEvent.to, swapEvent.from, options, pageTitleForBrowserHistory, historyState);
 	triggerEvent(TRANSITION_AFTER_SWAP);