+++ packages/db/src/core/integration/index.ts
@@ -1,20 +1,35 @@
 import { existsSync } from 'fs';
 import { dirname } from 'path';
 import { fileURLToPath } from 'url';
+import { LibsqlError } from '@libsql/client';
 import type { AstroConfig, AstroIntegration } from 'astro';
 import { mkdir, writeFile } from 'fs/promises';
 import { blue, yellow } from 'kleur/colors';
-import { loadEnv } from 'vite';
+import {
+	type HMRPayload,
+	type UserConfig,
+	type ViteDevServer,
+	createServer,
+	loadEnv,
+	mergeConfig,
+} from 'vite';
 import parseArgs from 'yargs-parser';
-import { SEED_DEV_FILE_NAME } from '../../runtime/queries.js';
 import { AstroDbError } from '../../runtime/utils.js';
 import { CONFIG_FILE_NAMES, DB_PATH } from '../consts.js';
+import { EXEC_DEFAULT_EXPORT_ERROR, EXEC_ERROR } from '../errors.js';
 import { resolveDbConfig } from '../load-file.js';
+import { SEED_DEV_FILE_NAME } from '../queries.js';
 import { type ManagedAppToken, getManagedAppTokenOrExit } from '../tokens.js';
 import { type VitePlugin, getDbDirectoryUrl } from '../utils.js';
 import { fileURLIntegration } from './file-url.js';
 import { typegenInternal } from './typegen.js';
-import { type LateSeedFiles, type LateTables, resolved, vitePluginDb } from './vite-plugin-db.js';
+import {
+	type LateSeedFiles,
+	type LateTables,
+	type SeedHandler,
+	resolved,
+	vitePluginDb,
+} from './vite-plugin-db.js';
 import { vitePluginInjectEnvTs } from './vite-plugin-inject-env-ts.js';
 
 function astroDBIntegration(): AstroIntegration {
@@ -22,6 +37,8 @@ function astroDBIntegration(): AstroIntegration {
 	let configFileDependencies: string[] = [];
 	let root: URL;
 	let appToken: ManagedAppToken | undefined;
+	// Used during production builds to load seed files.
+	let tempViteServer: ViteDevServer | undefined;
 
 	// Make table loading "late" to pass to plugins from `config:setup`,
 	// but load during `config:done` to wait for integrations to settle.
@@ -35,6 +52,13 @@ function astroDBIntegration(): AstroIntegration {
 			throw new Error('[astro:db] INTERNAL Seed files not loaded yet');
 		},
 	};
+	let seedHandler: SeedHandler = {
+		execute: () => {
+			throw new Error('[astro:db] INTERNAL Seed handler not loaded yet');
+		},
+		inProgress: false,
+	};
+
 	let command: 'dev' | 'build' | 'preview';
 	let output: AstroConfig['output'] = 'server';
 	return {
@@ -60,6 +84,7 @@ function astroDBIntegration(): AstroIntegration {
 						root: config.root,
 						srcDir: config.srcDir,
 						output: config.output,
+						seedHandler,
 					});
 				} else {
 					dbPlugin = vitePluginDb({
@@ -69,6 +94,8 @@ function astroDBIntegration(): AstroIntegration {
 						root: config.root,
 						srcDir: config.srcDir,
 						output: config.output,
+						logger,
+						seedHandler,
 					});
 				}
 
@@ -98,6 +125,9 @@ function astroDBIntegration(): AstroIntegration {
 				await typegenInternal({ tables: tables.get() ?? {}, root: config.root });
 			},
 			'astro:server:setup': async ({ server, logger }) => {
+				seedHandler.execute = async (fileUrl) => {
+					await executeSeedFile({ fileUrl, viteServer: server });
+				};
 				const filesToWatch = [
 					...CONFIG_FILE_NAMES.map((c) => new URL(c, getDbDirectoryUrl(root))),
 					...configFileDependencies.map((c) => new URL(c, root)),
@@ -118,46 +148,11 @@ function astroDBIntegration(): AstroIntegration {
 					const localSeedPaths = SEED_DEV_FILE_NAME.map(
 						(name) => new URL(name, getDbDirectoryUrl(root))
 					);
-					let seedInFlight = false;
-					// Load seed file on dev server startup.
+					// Eager load astro:db module on startup
 					if (seedFiles.get().length || localSeedPaths.find((path) => existsSync(path))) {
-						loadSeedModule();
-					}
-					const eagerReloadIntegrationSeedPaths = seedFiles
-						.get()
-						// Map integration seed paths to URLs, if possible.
-						// Module paths like `@example/seed` will be ignored
-						// from eager reloading.
-						.map((s) => (typeof s === 'string' && s.startsWith('.') ? new URL(s, root) : s))
-						.filter((s): s is URL => s instanceof URL);
-					const eagerReloadSeedPaths = [...eagerReloadIntegrationSeedPaths, ...localSeedPaths];
-					server.watcher.on('all', (event, relativeEntry) => {
-						if (event === 'unlink' || event === 'unlinkDir') return;
-						// When a seed file changes, load manually
-						// to track when seeding finishes and log a message.
-						const entry = new URL(relativeEntry, root);
-						if (eagerReloadSeedPaths.find((path) => entry.href === path.href)) {
-							loadSeedModule();
-						}
-					});
-
-					function loadSeedModule() {
-						if (seedInFlight) return;
-
-						seedInFlight = true;
-						const mod = server.moduleGraph.getModuleById(resolved.seedVirtual);
-						if (mod) server.moduleGraph.invalidateModule(mod);
-						server
-							.ssrLoadModule(resolved.seedVirtual)
-							.then(() => {
-								logger.info('Seeded database.');
-							})
-							.catch((e) => {
-								logger.error(e instanceof Error ? e.message : String(e));
-							})
-							.finally(() => {
-								seedInFlight = false;
-							});
+						server.ssrLoadModule(resolved.module).catch((e) => {
+							logger.error(e instanceof Error ? e.message : String(e));
+						});
 					}
 				}, 100);
 			},
@@ -175,8 +170,15 @@ function astroDBIntegration(): AstroIntegration {
 
 				logger.info('database: ' + (connectToStudio ? yellow('remote') : blue('local database.')));
 			},
+			'astro:build:setup': async ({ vite }) => {
+				tempViteServer = await getTempViteServer({ viteConfig: vite });
+				seedHandler.execute = async (fileUrl) => {
+					await executeSeedFile({ fileUrl, viteServer: tempViteServer! });
+				};
+			},
 			'astro:build:done': async ({}) => {
 				await appToken?.destroy();
+				await tempViteServer?.close();
 			},
 		},
 	};
@@ -189,4 +191,49 @@ function databaseFileEnvDefined() {
 
 export function integration(): AstroIntegration[] {
 	return [astroDBIntegration(), fileURLIntegration()];
+}
+
+async function executeSeedFile({
+	fileUrl,
+	viteServer,
+}: {
+	fileUrl: URL;
+	viteServer: ViteDevServer;
+}) {
+	const mod = await viteServer.ssrLoadModule(fileUrl.pathname);
+	if (typeof mod.default !== 'function') {
+		throw new AstroDbError(EXEC_DEFAULT_EXPORT_ERROR(fileURLToPath(fileUrl)));
+	}
+	try {
+		await mod.default();
+	} catch (e) {
+		if (e instanceof LibsqlError) {
+			throw new AstroDbError(EXEC_ERROR(e.message));
+		}
+		throw e;
+	}
+}
+
+/**
+ * Inspired by Astro content collection config loader.
+ */
+async function getTempViteServer({ viteConfig }: { viteConfig: UserConfig }) {
+	const tempViteServer = await createServer(
+		mergeConfig(viteConfig, {
+			server: { middlewareMode: true, hmr: false, watch: null },
+			optimizeDeps: { noDiscovery: true },
+			ssr: { external: [] },
+			logLevel: 'silent',
+		})
+	);
+
+	const hotSend = tempViteServer.hot.send;
+	tempViteServer.hot.send = (payload: HMRPayload) => {
+		if (payload.type === 'error') {
+			throw payload.err;
+		}
+		return hotSend(payload);
+	};
+
+	return tempViteServer;
 }