+++ packages/integrations/mdx/test/fixtures/mdx-optimize/astro.config.mjs
@@ -3,7 +3,7 @@ import mdx from '@astrojs/mdx';
 export default {
 	integrations: [mdx({
 		optimize: {
-			customComponentNames: ['strong']
+			ignoreElementNames: ['strong']
 		}
 	})]
 }