+++ packages/integrations/vercel/src/serverless/entrypoint.ts
@@ -12,7 +12,7 @@ applyPolyfills();
 
 export const createExports = (
 	manifest: SSRManifest,
-	{ middlewareSecret }: { middlewareSecret: string }
+	{ middlewareSecret, skewProtection }: { middlewareSecret: string; skewProtection: boolean }
 ) => {
 	const app = new NodeApp(manifest);
 	const handler = async (req: IncomingMessage, res: ServerResponse) => {
@@ -37,6 +37,11 @@ export const createExports = (
 		}
 		// hide the secret from the rest of user code
 		delete req.headers[ASTRO_MIDDLEWARE_SECRET_HEADER];
+
+		// https://vercel.com/docs/deployments/skew-protection#supported-frameworks
+		if (skewProtection && process.env.VERCEL_SKEW_PROTECTION_ENABLED === '1') {
+			req.headers['x-deployment-id'] = process.env.VERCEL_DEPLOYMENT_ID;
+		}
 
 		const webResponse = await app.render(req, { addCookieHeader: true, clientAddress, locals });
 		await NodeApp.writeResponse(webResponse, res);